feat: JWT 인증/인가, 직원·부서 CRUD, 페이지네이션 구현
- Spring Security + JWT(jjwt 0.12.6) 필터 체인 구성
  - JwtAuthenticationFilter: 토큰 검증 및 SecurityContext 설정
  - JwtTokenProvider: Access(30분) / Refresh(7일) 토큰 발급·검증
  - JwtAuthenticationEntryPoint / JwtAccessDeniedHandler: 401·403 JSON 응답
  - LocalSecurityConfig: local 프로파일 전용 H2 콘솔 허용 체인
- 인증 API: 로그인, 로그아웃(Redis Refresh 토큰 단일 세션), 토큰 재발급
- 직원 CRUD: 목록(페이지네이션+필터), 등록, 상세 조회, 수정, 퇴사 처리, 내 정보 조회·수정
- 부서 CRUD: 목록(N+1 방지 집계 쿼리), 등록, 수정, 삭제
- 공통 인프라: ApiResponse, PageResponse, ErrorCode, GlobalExceptionHandler
- Flyway V1 전체 스키마(14테이블) + V2 시드 데이터 마이그레이션
- 보안 강화: docker-compose Redis 포트 localhost 바인딩, 테스트 stdout 비활성화

+++ src/main/java/com/grown/smartoffice/domain/department/dto/DepartmentListResponse.java
@@ -13,7 +13,7 @@ public class DepartmentListResponse {
     private Long id;
     private String name;
     private String description;
-    private int userCount;
+    private long userCount;
     private LocalDateTime createdAt;
 
     public static DepartmentListResponse of(Department department, long userCount) {
@@ -21,7 +21,7 @@ public class DepartmentListResponse {
                 .id(department.getDeptId())
                 .name(department.getDeptName())
                 .description(department.getDeptDescription())
-                .userCount((int) userCount)
+                .userCount(userCount)
                 .createdAt(department.getCreatedAt())
                 .build();
     }
